requirements.txt
fabric
coverage
django~=2.2
markdown
gunicorn
faker
django-pure-pagination
elasticsearch>=2,<3
django-haystack
djangorestframework
django-filter
mysqlclient==2.0.3
django-cors-headers